// src/app/agendamento/model/Evento.ts
export interface Evento {
    id_evento: number;
    descricao_evento: string;
    observacao_evento: string;
    inicio: Date;
    termino: Date;
    agenda: string;
    cliente: string;
    animal: string;
    logradouro: string
    telefone: string;
    celular: string;
    bairro: string;
    cidade: string;
    uf: string;
    atendente: string;
    empresa: string;
}